Relicense under GPLv3
Hereby, the plugin code in this project is re-released under the terms
of the GNU Public License, version 3. This applies to the plugin code,
all of which in its current form was written by Reuben Vinal. After
contacting him, he gave me permission to re-release all the code found
on [1] under the terms of the GPLv3.

Note that this does not apply to the third-party code that is
currently committed to this repository:
 - CVSTHost by Herrmann Seib is already licensed under GPLv2 or later.
 - The Reaper SDK has its own, custom license.
 - JUCE code has its own license, notably the currently committed
   version is compatible with GPLv2. All JUCE source code will soon
   be replaced by the current version of JUCE which is compatible to
   GPLv2 but not GPLv3.

Dedicated copyright headers for each single file will be added soon.

[1]: https://code.google.com/archive/p/pizmidi/
     Latest commit hash: c4352c7d1c2964c0e2dc03e408cb35f6e9e5247b.

Signed-off-by: Simon Leiner <simon@leiner.me>