Achieve compatibility with CMake 3.21
Compatibility with CMake 3.21 is very convient as this is the version
bundled with the current version of Visual Studio 2022, namely 17.0.4.

The only feature from 3.22 used up until now is the _environment
variable_ for CMAKE_BUILD_TYPE. However, the build type can still be
set from the presets by setting it as a cache variable instead.

Signed-off-by: Simon Leiner <simon@leiner.me>