CMake: Use multi-config generators, add build presets
Using build presets streamlines the invocation of the CMake build tool
command and allows for uniform indication of the build type across all
operating systems: On Windows, typically Visual Studio is used as
generator. Visual Studio inherently is a multi-config generator. So, in
switching to the "Ninja Multi-Config" generator for macOS and Unix,
this patch lets the build type be specified at build invocation (rather
than the configure step) for all operating systems.

Signed-off-by: Simon Leiner <simon@leiner.me>