Fix curl linkage problems on Ubuntu
Similar to JUCE_WEB_BROWSER and JUCE_NEEDS_WEB_BROWSER, JUCE_USE_CURL
and JUCE_NEEDS_CURL need to be set consistently in order to avoid
linker problems.
However, in contrast to JUCE_NEEDS_WEB_BROWSER, JUCE_NEEDS_CURL must
be given to juce_add_plugin() instead of adding it as a target
property after the call. Because of that, this commit modifies the
structure of piz_add_plugin().

Signed-off-by: Simon Leiner <simon@leiner.me>